requirements.txt
Flask
pillow
numpy
sap_cf_logging